requirements.txt
streamlit
langchain
faiss-cpu==1.7.4
pypdf
sentence-transformers
google-generativeai
python-dotenv
gtts
langchain-community
langchain-core
httpx==0.27.0
graphviz